Fixed some Graphic & core issues found in marshmallow
Major Rework for GUI for android versions 23+;

Fixed issue in regard to nexus where a closed case called onStop()
method to be called and stop the alarm. Seeing as on back press, on
StopAlarm & onPauseAlarm all call the onFinish() method the onDestroy()
was overridden and the stop alarm function was called there.
Added a few optimisations